requirements.txt
#LangChain Core
langchain 
langchain-core 

#Open-AI Integration
langchain-openai
openai

#Anthropic integration
langchain-anthropic 

#Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Environment Variables Management
python-dotenv

#Machine Learning utilities
numpy
scikit-learn 

pytorch